Allow specifying serial or structure for away
Adds the global -S/--structure argument for specifying the structure
name. Away status will be scoped to the given structure name or to the
structure that a device appears in, falling back to just the first
structure found.

Also scopes device discovery for all other functions if specifed to only
the devices in that structure.

Fixes #8