Added some more useful information for Hot Water functionality
In the UK and Europe, the third generation Nest Thermostat has hot
water on/off control functionality. These properties give information
on this including if the hot water is on or off (hot_water_active) and
if it is on a boost, then when the boost time ends
(hot_water_boost_time_to_end).