Add way to check if (re-)autorization is needed
Previously, the only way to check this was to check if there's an access
token. That only cathes cases where user never authorized. It fails to
account for a user disconnecting the integration or the token otherwise
expiring.

In addition, it's possible to update the nest product's permissions and
client not to have them.

Now we have 'authorization_required' to check all these cases. The
latter only happens if you specify a product_version. Alternatively,
there are 3 helper methods to check for each state in case you want to
have your own logic.