Handle missing where names
Due to an upstream bug some devices will be assigned a where_id that is not
visible in the API. As a result we can't get a friendly name for the where.

If we encounter a where_id with no associated name just return the id as
the name.

See https://nestdevelopers.io/t/missing-where-name-from-some-devices/1202

Fixes #127